delete y cotizacion

+++ src/repositories/permiso.repository.js
@@ -1,4 +1,4 @@
-const { Permiso } = require("../models");
+const { Permiso, RolesPermisos } = require("../models");
 
 exports.getAllPermisos = async () => {
     return await Permiso.findAll();
@@ -20,5 +20,22 @@ exports.statusPermiso = async (id) => {
     return await Permiso.update({ estado: false }, { where: { id } });
 }
 exports.deletePermiso = async (id) => {
-    return await Permiso.destroy( { where: { id } });
+    const permiso = await Permiso.findByPk(id);
+    
+    if (!permiso) {
+        throw new Error("Permiso no encontrado");
+    }
+
+    const existeEnRolesPermisos = await RolesPermisos.findOne({ where: { permisoId: id } });
+    
+    if (existeEnRolesPermisos) {
+        throw new Error("No se puede eliminar el permiso porque está asociado a un rol");
+    }
+
+    if (permiso.estadoId === 2) {
+        return await Permiso.destroy( { where: { id } });
+    }
+    else {
+        throw new Error("No se puede eliminar el permiso porque está activo");
+    }
 }
